Change format of new node resonse, and refactor node primary key
