+++ src/components/shared/searchView/NodeItem.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { IoIosRadioButtonOn } from 'react-icons/io';
-import INodeBase from '~/models/baseModels/INodeBase';
 import routeBuilder from '~/routing/routeBuilder';
 import NodeTypeHelper from '~/util/nodeTypeHelper';
 import SubSites from '~/routing/subSites';
 import navigator from '~/routing/navigator';
+import { INodeBase } from '~/models/INode';
 import * as s from './nodeItem.scss';
 
 interface INodeItemProps {
