+++ src/stores/searchResultStore.ts
@@ -1,6 +1,6 @@
 import { action, computed, observable, reaction } from 'mobx';
 import { ILine } from '~/models';
-import INodeBase from '~/models/baseModels/INodeBase';
+import { INodeBase } from '~/models/INode';
 import TransitType from '~/enums/transitType';
 import SearchStore from './searchStore';
